server/requirements.txt
alembic==0.9.6
aniso8601==1.3.0
astroid==1.6.5
Babel==2.6.0
bcrypt==3.1.4
cairocffi==0.9.0
CairoSVG==2.1.3
cffi==1.11.2
click==6.7
colorama==0.3.9
coverage==4.4.2
cssselect2==0.2.1
decorator==4.3.0
defusedxml==0.5.0
enum-compat==0.0.2
eventlet==0.21.0
Flask==0.12.2
Flask-AppBuilder==1.11.1
Flask-Babel==0.11.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.6
Flask-Login==0.2.11
Flask-Migrate==2.1.1
Flask-OpenID==1.2.5
flask-restplus==0.10.1
Flask-Script==2.0.6
Flask-SocketIO==2.9.3
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-WeasyPrint==0.5
Flask-WTF==0.14.2
gem==0.1.12
greenlet==0.4.12
gunicorn==19.7.1
html5lib==1.0.1
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
pdfdocument==3.2
pdfrw==0.4
Pillow==5.2.0
pkg-resources==0.0.0
psycopg2==2.7.3.2
py==1.5.4
pycparser==2.18
PyJWT==1.5.3
pylint==1.8.2
pymongo==3.3.1
Pyphen==0.9.5
pytest==3.0.6
python-dateutil==2.6.1
python-editor==1.0.3
python-engineio==2.0.2
python-socketio==1.8.4
python3-openid==3.1.0
pytz==2017.3
reportlab==3.5.6
selenium==3.8.1
six==1.11.0
SQLAlchemy==1.2.0
sqlalchemy-migrate==0.7.2
TableFactory==0.1.4
Tempita==0.5.2
tinycss2==0.6.1
WeasyPrint==0.42.3
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
xlwt==1.3.0
